Fixed with comments

+++ src/main/java/ru/practicum/shareit/user/UserServiceImpl.java
@@ -65,6 +65,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(Long id) {
+        if (!userRepository.findById(id).isPresent()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Пользователь не найден");
+        }
         userRepository.deleteById(id);
     }
 }
